+++ canopy/forms/__init__.py
@@ -1,9 +1,10 @@
-from .slot import SlotAdminForm
+from .slot import SlotAdminForm, SlotAdminInlineForm
 from .options import build_options_form, BaseSlotOptionsForm
 
 
 __all__ = [
     "SlotAdminForm",
+    "SlotAdminInlineForm",
     "BaseSlotOptionsForm",
     "build_options_form",
 ]